+++ src/app/snackbar/snackbar.sample.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, ViewChild } from '@angular/core';
 import { IgxSnackbarComponent } from 'igniteui-angular';
 
 @Component({
@@ -7,8 +7,10 @@ import { IgxSnackbarComponent } from 'igniteui-angular';
     templateUrl: 'snackbar.sample.html'
 })
 export class SnackbarSampleComponent implements OnInit {
+    @ViewChild('snackbar')
+    private snackbar: IgxSnackbarComponent;
+
     public color: string;
-    public message: string;
     public actionName: string;
     private _colors: string[];
 
@@ -29,13 +31,24 @@ export class SnackbarSampleComponent implements OnInit {
         this._colors.push(this.color);
         this.color = color;
 
-        this.message = 'Changed color to ' + this.color;
-        snackbar.open();
+        snackbar.open('Changed color to ' + this.color);
     }
 
     public undoColorChange(snackbar) {
         this.color = this._colors.pop();
 
         snackbar.close();
+    }
+
+    public onAnimationStarted() {
+        console.log('animation started');
+    }
+
+    public onAnimationDone() {
+        console.log('animation ended');
+    }
+
+    public toggleSnackbar() {
+        this.snackbar.toggle();
     }
 }